+++ src/CleanRegex/Internal/EntryPoints.php
@@ -86,11 +86,6 @@ trait EntryPoints
         return new PatternBuilder();
     }
 
-    public static function quote(string $string): string
-    {
-        return preg::quote($string);
-    }
-
     public static function unquote(string $quotedString): string
     {
         return (new UnquotePattern($quotedString))->unquote();